+++ src/library/UniformResourceIdentifier/definition.ts
